renaming brca_pseudonym_generator_new brca_pseudonym_generator
